% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_european_choropleth.R
\name{get_european_choropleth}
\alias{get_european_choropleth}
\title{Prepare a ggplot2-based choropleth on European national (NUTS0) countries.}
\usage{
get_european_choropleth(input_data)
}
\arguments{
\item{input_data}{A data frame containing the data to be mapped. It should contain two columns: the first containing the ISO ALPHA-2
code of the country, the second the value that should be mapped}
}
\value{
A ggplot2 object
}
\description{
Prepare a ggplot2-based choropleth on European national (NUTS0) countries.
}
