% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_region_from_coordinates.R
\name{get_region_from_coordinates}
\alias{get_region_from_coordinates}
\title{Return the geographci region of a set of coordinates according to administrative boundaries.}
\usage{
get_region_from_coordinates(obj, shapefile = "NUTS0",
  path_to_shapefile = NULL)
}
\arguments{
\item{obj}{A data frame of coordinates stored using lat/lon coordinates. See Details section for further details.}

\item{shapefile}{The shapefile to be used to aggregate the grid points.}
}
\value{
A vector containing all the \code{shapefile_id_field}
}
\description{
The function defines in which region a set of coordinates (stored into a \code{data.frame}) lie
The administrative boundaries can be chosen from a set of pre-definites shapefiles.
}
\details{
Details
\code{obj} must be a data frame containing two variables: 
1. Latitude: a double in the range -90, 90 named as lat, Lat, latitude, latit, etc.
2. Longitude: a double in the range -180, 180 named as lon, Lon, longitude, long, etc.

The shapefiles available are the following:
\itemize{
\item \code{NUTS0-2}: Data from EUROSTAT NUTS (small islands removed) https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units/nutscountries_EU 
\item \code{eh2050}: cluster as defined in the FP7 e-Highway2050 project
\item \code{hybas05}: HydroBASINS Level 5 (http://www.hydrosheds.org/page/hydrobasins)
\item \code{hybas06}: HydroBASINS Level 6 (http://www.hydrosheds.org/page/hydrobasins)
\item \code{WAPP}: WAPP catchments from the JRC LISFLOOD hydrological model
}
}
\author{
Matteo De Felice
}
