% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_field_discrete.R
\name{plot_field_discrete}
\alias{plot_field_discrete}
\title{Plot a gridded field using discrete scale}
\usage{
plot_field_discrete(x, lon, lat, lonlim = "auto", latlim = "auto",
  labels = c(), breaks = c(), cscale = "Spectral", rev_cscale = F,
  show_lat_lon_labels = T, varname = "x", title = c(), mask = NULL,
  siglev = NULL, plot_only_sig = F, smooth = F, smooth_factor = 5,
  smooth_theta = 0.5, lineWidth = 0.5, dotSize = 0.5, GRID_STEP = 10,
  FONT_SIZE = 18, show_legend = T)
}
\arguments{
\item{x}{The field to be plotted, it can be both a matrix or a data frame. See Details for more information.}

\item{lon}{Vector of longitude values}

\item{lat}{Vector of latitude values}

\item{lonlim}{A vector of two values defining the longitude limits for the plot. The default is 'auto': it tries to detect the extreme directly from the provided longitude values.}

\item{òatlim}{A vector of two values defining the latitude limits for the plot. The default is 'auto': it tries to detect the extreme directly from the provided latitude values.}
}
\description{
This function tries to produce high-quality maps of gridded fields using a discrete color-scale. The function is based on \code{ggplot2}.
}
\details{
Details
About x.
}
