% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_labels.R
\name{pretty_labels}
\alias{pretty_labels}
\title{Create pretty labels useful to visualise discrete intervals}
\usage{
pretty_labels(breaks, separator_character = ",")
}
\arguments{
\item{breaks}{Vector of breaks' values}

\item{separator_character}{Character that is used to separate the values of the range shown in the legend. Default is ",".}
}
\value{
A caharacter vector with all the labels. The labels' lenght is equals to the length of the provided breaks +2, the extremes.
}
\description{
The function creates pretty labels for the breaks used in discrete scales in \code{ggplot2}
}
\author{
Matteo De Felice
}
