% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ts_from_shp.R
\name{get_ts_from_shp}
\alias{get_ts_from_shp}
\title{Aggregate a gridded object into national or regional boundaries.}
\usage{
get_ts_from_shp(obj, weight_matrix = NULL, aggregate_function = "mean",
  shapefile = "NUTS0", shapefile_id_field = "NUTS_ID",
  path_to_shapefile = NULL, cos_weighted = TRUE)
}
\arguments{
\item{obj}{A gridded object: a grid structure from Climate4R functions or a \code{list} with \code{lat}, \code{lon} and \code{data} fields. See Details section for further details.}

\item{weight_matrix}{A matrix to be used as weight when aggregating}

\item{aggregate_function}{The function to be used to aggregate the grid points. Options are: \code{mean} and \code{sum}}

\item{shapefile}{The shapefile to be used to aggregate the grid points. The options are: \code{NUTS0}, \code{NUTS1}, \code{NUTS2}, \code{eh2050} and \code{custom}}

\item{shapefile_id_field}{Field name of the shapefile to be used as identifier}

\item{path_to_shapefile}{The path of the \code{.shp} file to be used when \code{shapefile} is set to \code{custom}}

\item{cos_weighted}{Define is the grid points will be weighted according the cosine of the latitude}
}
\value{
A list containing all the fields names ad \code{shapefile_id_field}, each field contains aggregated the time-series
}
\description{
The function spatially aggregate using a function (for example, \code{mean}) all the grid points
within the borders defined by a shapefile. The function already provides some shapefiles for
Europe.
}
\details{
Details
The \code{obj} can be: 1. a list with three mandatory fields: \code{lat} with the latitude values, \code{lon} with the longitude and \code{data} with the gridded field consistent with the coordinates;
2. a grid structure as in the \code{Climate4R} bundle (citation neeeded), for example from \code{loadeR} package.

The shapefiles available are the following:
- \code{NUTS0}: countries_EU (specify source)
- \code{NUTS1}: NUTS_REG_01M_2013_ADM1 (specify source)
- \code{NUTS2}: NUTS_REG_01M_2013_REGIONS (specify source)
- \code{eh2050}: cluster as defined in the FP7 e-Highway2050 project
}
\author{
Matteo De Felice
}
